Initial commit
This is the initial commit of ferro, a library to enable writing "playbooks"
in Rust, in the style of Ansible.

A playbook consists of a vector of Tasks, each of which defines a module to
run, and its parameters. The playbook is given a HashMap of "vars", and some
initial state. The playbook registers the output of task modules in the state
as it runs. Unlike Ansible, which requires the user to "register" a task's
output into a variable, Ferro does it automatically for all tasks, storing it
as in state using the task's description as a key. Subsequent tasks can get
the output of a previous task by calling the lazy::state() function. Tasks can
also get the value of vars with the lazy::var() function.

Three modules are included:
* NullModule - A module that does nothing.
* Command - A module to run commands.
* CloudFormation - A module to manage CloudFormation stacks.