Make CloudFormation template a lazy value
All module parameters should be lazy so they can be determined at runtime
based on the outputs of other modules. This converts the CloudFormation
template parameter to a function taking a Context, as other parameters do.
The Context has access to the variables and state of the playbook so
values can be retrieved from them at runtime.