Make all modules implement Default, and remove new() methods
All modules will be created as literal structs in playbooks instead of with
a new() method. Default::default() will be called automatically for every
module during the instantiation of the struct. This makes implementation
easier, although it has the downside that the module struct cannot contain
any private members. This may be revisited later when more of the pieces are
in place.